Use tmp dir in HTMLPurifier
