+++ tests/PhpSpreadsheetTests/Writer/Html/InvalidFileNameTest.php
@@ -39,7 +39,7 @@ class InvalidFileNameTest extends Functional\AbstractFunctional
         self::assertEquals('Helvetica', $writer->getFont());
         $writer->setPaperSize(PageSetup::PAPERSIZE_LEDGER);
         self::assertEquals($writer->getPaperSize(), PageSetup::PAPERSIZE_LEDGER);
-        self::assertEquals(File::sysGetTempDir() . '/phpsp/pdf', $writer->getTempDir());
+        self::assertEquals(File::sysGetTempDir() . '/phpsppdf', $writer->getTempDir());
         $writer->setTempDir(File::sysGetTempDir());
         self::assertEquals(File::sysGetTempDir(), $writer->getTempDir());
     }
